🌈 style(UserSection): ux/ui user section
Implemented user external link to go to user github page and improvement desgin in UserSection
