+++ src/components/user/types.ts
@@ -1,0 +1,11 @@
+export interface User {
+    avatar_url: string;
+    login: string;
+    name: string;
+    location: string;
+    bio: string;
+    public_repos: number;
+    followers: number;
+    following: number;
+    html_url: string;
+}
